Pantheon/Calculator_Oracle:ESG--GSG__PROD@v0.1.0/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.4
pytest==7.4.4
httpx==0.26.0
python-multipart==0.0.6